services/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Poetry init for server service"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-socketio (>=5.5.1,<6.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-bcrypt (>=1.0.1,<2.0.0)",
    "flask-cors (>=6.0.2,<7.0.0)",
    "pymysql (>=1.1.2,<2.0.0)",
    "cryptography (>=46.0.3,<47.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
